backend/requirements.txt
django
djangorestframework
requests
beautifulsoup4
django-cors-headers
gunicorn
python-dotenv
